backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-multipart
pymupdf
beautifulsoup4
requests
openai>=1.0.0
jinja2
python-dotenv
